some Fixes:   PlayerLogin.cs -> 63 -> Zeilenumbruch entfernt   StorePlayerData.cs -> 64 -> (decimal)player.getSyncedData("Money") to (float)(decimal)player.getSyncedData("Money")   StorePlayerData.cs -> 64 -> (decimal)player.getSyncedData("BankAccount") to (float)(decimal)player.getSyncedData("BankAccount")

+++ resources/TerraTex-RL-RPG/Lib/Threads/StorePlayerData.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading;
 using GrandTheftMultiplayer.Server.Elements;
+using GrandTheftMultiplayer.Server.API;
 using MySql.Data.MySqlClient;
 
 namespace TerraTex_RL_RPG.Lib.Threads
@@ -61,8 +62,11 @@ namespace TerraTex_RL_RPG.Lib.Threads
                                             "Phone = @Phone " +
                                             "WHERE UserID = @UserID";
 
-            updateUserCommand.Parameters.AddWithValue("@Money", (decimal)player.getSyncedData("Money"));
-            updateUserCommand.Parameters.AddWithValue("@BankAccount", (decimal)player.getSyncedData("BankAccount"));
+
+
+            updateUserCommand.Parameters.AddWithValue("@Money", (float)(decimal)player.getSyncedData("Money"));
+            updateUserCommand.Parameters.AddWithValue("@BankAccount", (float)(decimal)player.getSyncedData("BankAccount"));
+
             if (player.getSyncedData("Phone") != -1)
             {
                 updateUserCommand.Parameters.AddWithValue("@Phone", player.getSyncedData("Phone"));
